+++ src/Lib.hs
@@ -7,170 +7,59 @@
 
 {- Exports -}
 module Lib
-    ( parseRomeOptions
-    , runRomeWithOptions
+    ( runRomeWithOptions
     , discoverRegion
     , filterByNameEqualTo
     , filterOutFrameworkNamesAndVersionsIfNotIn
-    , splitWithSeparator
     ) where
 
 
 
 {- Imports -}
 import qualified Codec.Archive.Zip            as Zip
+import           Configuration
 import           Control.Applicative          ((<|>))
 import           Control.Lens                 hiding (List)
 import           Control.Monad
 import           Control.Monad.Except
-import           Control.Monad.Reader         (MonadReader, ReaderT, ask,
-                                               runReaderT)
+import           Control.Monad.Reader         (ReaderT, ask, runReaderT)
 import           Control.Monad.Trans          (MonadIO, lift, liftIO)
 import           Control.Monad.Trans.Maybe
 import           Control.Monad.Trans.Resource (runResourceT)
 import qualified Data.ByteString              as BS
 import qualified Data.ByteString.Lazy         as LBS
 import           Data.Cartfile
-import           Data.Char                    (isLetter)
-import qualified Data.Conduit                 as C (Conduit, Sink, await, yield, ($$),
-                                                    (=$=))
+import qualified Data.Conduit                 as C (Conduit, Sink, await, yield,
+                                                    ($$), (=$=))
 import qualified Data.Conduit.Binary          as C (sinkFile, sinkLbs,
                                                     sourceFile, sourceLbs)
-import           Data.Either.Utils
-import           Data.Function
-import           Data.GitRepoAvailability
+import           Data.Function                (on)
 import           Data.Ini                     as INI
 import           Data.Ini.Utils               as INI
 import           Data.List
-import           Data.List.Split
-import qualified Data.Map.Strict              as M
-import           Data.Maybe
+import           Data.Maybe                   (fromMaybe)
+import           Data.Monoid                  ((<>))
 import           Data.Romefile
-import           Data.String.Utils
-import           Data.TargetPlatform
 import qualified Data.Text                    as T
-import           Data.Time
 import qualified Network.AWS                  as AWS
 import           Network.AWS.Data
 import           Network.AWS.S3               as S3
-import           Options.Applicative          as Opts
 import           System.Directory
 import           System.Environment
 import           System.FilePath
-import           Text.Read
-
-{- Types -}
-
-type UDCEnv                = (AWS.Env{-, VerifyFlag-}, SkipLocalCacheFlag, Bool)
-type RomeMonad             = ExceptT String IO
-type RepositoryMap         = M.Map GitRepoName [FrameworkName]
-type InvertedRepositoryMap = M.Map FrameworkName GitRepoName
-
-data RomeCommand = Upload RomeUDCPayload
-                  | Download RomeUDCPayload
-                  | List RomeListPayload
-                  deriving (Show, Eq)
-
-data RomeUDCPayload = RomeUDCPayload { _payload            :: [GitRepoName]
-                                     , _udcPlatforms       :: [TargetPlatform]
-                                    --  , _verifyFlag         :: VerifyFlag
-                                     , _skipLocalCacheFlag :: SkipLocalCacheFlag
-                                     }
-                                     deriving (Show, Eq)
-
--- newtype VerifyFlag = VerifyFlag { _verify :: Bool } deriving (Show, Eq)
-newtype SkipLocalCacheFlag = SkipLocalCacheFlag { _skipLocalCache :: Bool } deriving (Show, Eq)
-
-data RomeListPayload = RomeListPayload { _listMode      :: ListMode
-                                       , _listPlatforms :: [TargetPlatform]
-                                       }
-                                       deriving (Show, Eq)
-
-data ListMode = All
-               | Missing
-               | Present
-               deriving (Show, Eq)
-
-data RomeOptions = RomeOptions { romeCommand :: RomeCommand
-                               , verbose     :: Bool
-                               }
+import           Types
+import           Types.Commands               as Commands
+import           Types.TargetPlatform
+import           Utils
 
 
-{- Constants -}
-carthageBuildDirectory :: TargetPlatform -> FilePath
-carthageBuildDirectory platform = "Carthage" </> "Build" </> show platform
-
-{- Commnad line arguments parsing -}
-
--- verifyParser :: Parser VerifyFlag
--- verifyParser = VerifyFlag <$> Opts.switch ( Opts.long "verify" <> Opts.help "Verify that the framework has the same hash as specified in the Cartfile.resolved.")
-
-skipLocalCacheParser :: Parser SkipLocalCacheFlag
-skipLocalCacheParser = SkipLocalCacheFlag <$> Opts.switch ( Opts.long "skip-local-cache" <> Opts.help "Ignore the local cache when performing the operation.")
-
-reposParser :: Opts.Parser [GitRepoName]
-reposParser = Opts.many (Opts.argument (GitRepoName <$> str) (Opts.metavar "FRAMEWORKS..." <> Opts.help "Zero or more framework names. If zero, all frameworks and dSYMs are uploaded."))
-
-platformsParser :: Opts.Parser [TargetPlatform]
-platformsParser = (nub . concat <$> Opts.some (Opts.option (eitherReader platformListOrError) (Opts.metavar "PLATFORMS" <> Opts.long "platform" <> Opts.help "Applicable platforms for the command. One of iOS, MacOS, tvOS, watchOS, or a comma-separated list of any of these values.")))
-  <|> pure allTargetPlatforms
-  where
-    platformOrError str = maybeToEither ("Unrecognized platform '" ++ str ++ "'") (readMaybe str)
-    splitPlatforms str = filter (not . null) $ filter isLetter <$> wordsBy (not . isLetter) str
-    platformListOrError str = mapM platformOrError $ splitPlatforms str
-
-udcPayloadParser :: Opts.Parser RomeUDCPayload
-udcPayloadParser = RomeUDCPayload <$> reposParser <*> platformsParser {- <*> verifyParser-} <*> skipLocalCacheParser
-
-uploadParser :: Opts.Parser RomeCommand
-uploadParser = pure Upload <*> udcPayloadParser
-
-downloadParser :: Opts.Parser RomeCommand
-downloadParser = pure Download <*> udcPayloadParser
-
-listModeParser :: Opts.Parser ListMode
-listModeParser = (
-                    Opts.flag' Missing (Opts.long "missing" <> Opts.help "List frameworks missing from the cache. Ignores dSYMs")
-                    <|> Opts.flag' Present (Opts.long "present" <> Opts.help "List frameworks present in the cache. Ignores dSYMs.")
-                 )
-                <|> Opts.flag All All (Opts.help "Reports missing or present status of frameworks in the cache. Ignores dSYMs.")
-
-listPayloadParser :: Opts.Parser RomeListPayload
-listPayloadParser = RomeListPayload <$> listModeParser <*> platformsParser
-
-listParser :: Opts.Parser RomeCommand
-listParser = List <$> listPayloadParser
-
-parseRomeCommand :: Opts.Parser RomeCommand
-parseRomeCommand = Opts.subparser $
-  Opts.command "upload" (uploadParser `withInfo` "Uploads frameworks and dSYMs contained in the local Carthage/Build/<platform> to S3, according to the local Cartfile.resolved")
-  <> Opts.command "download" (downloadParser `withInfo` "Downloads and unpacks in Carthage/Build/<platform> frameworks and dSYMs found in S3, according to the local Carftfile.resolved")
-  <> Opts.command "list" (listParser `withInfo` "Lists frameworks in the cache and reports cache misses/hits, according to the local Carftfile.resolved. Ignores dSYMs.")
-
-parseRomeOptions :: Opts.Parser RomeOptions
-parseRomeOptions = RomeOptions <$> parseRomeCommand <*> Opts.switch ( Opts.short 'v' <> help "Show verbose output" )
-
-withInfo :: Opts.Parser a -> String -> Opts.ParserInfo a
-withInfo opts desc = Opts.info (Opts.helper <*> opts) $ Opts.progDesc desc
-
-{- Functions -}
-
-getCartfileEntires :: RomeMonad [CartfileEntry]
-getCartfileEntires = do
-  eitherCartfileEntries <- liftIO $ parseCartfileResolved cartfileResolved
-  case eitherCartfileEntries of
-    Left e -> throwError $ "Carfile.resolved parse error: " ++ show e
-    Right cartfileEntries -> return cartfileEntries
-
-getRomefileEntries :: RomeMonad RomeFileParseResult
-getRomefileEntries = parseRomefile romefile
 
 runRomeWithOptions :: AWS.Env -> RomeOptions -> RomeMonad ()
 runRomeWithOptions env (RomeOptions options verbose) = do
   cartfileEntries <- getCartfileEntires
   RomeFileParseResult { .. } <- getRomefileEntries
-  let respositoryMap = toRomeFilesEntriesMap repositoryMapEntries
-  let reverseRepositoryMap = toInvertedRomeFilesEntriesMap repositoryMapEntries
+  let respositoryMap = toRepositoryMap repositoryMapEntries
+  let reverseRepositoryMap = toInvertedRepositoryMap repositoryMapEntries
   let ignoreNames = concatMap frameworkCommonNames ignoreMapEntries
   case options of
 
@@ -196,25 +85,6 @@ runRomeWithOptions env (RomeOptions options verbose) = do
         let repoAvailabilities = getMergedGitRepoAvailabilitiesFromFrameworkAvailabilities reverseRepositoryMap availabilities
         let repoLines = filter (not . null) $ fmap (formattedRepoAvailability listMode) repoAvailabilities
         mapM_ sayLn repoLines
-
-  where
-
-    constructFrameworksAndVersionsFrom :: [CartfileEntry] -> RepositoryMap -> [FrameworkVersion]
-    constructFrameworksAndVersionsFrom cartfileEntries repositoryMap = deriveFrameworkNamesAndVersion repositoryMap cartfileEntries
-    filterRepoMapByGitRepoNames :: RepositoryMap -> [GitRepoName] -> RepositoryMap
-    filterRepoMapByGitRepoNames repoMap gitRepoNames = M.unions $ map (restrictRepositoryMapToGitRepoName repoMap) gitRepoNames
-
-fromErrorMessage :: AWS.ErrorMessage -> String
-fromErrorMessage (AWS.ErrorMessage t) = T.unpack t
-
-filterByNameEqualTo :: [FrameworkVersion] -> FrameworkName -> [FrameworkVersion]
-filterByNameEqualTo fs s = filter (\(FrameworkVersion name version) -> name == s) fs
-
-filterOutFrameworkNamesAndVersionsIfNotIn :: [FrameworkVersion] -> [FrameworkName] -> [FrameworkVersion]
-filterOutFrameworkNamesAndVersionsIfNotIn favs fns = [fv |  fv <- favs,  _frameworkName fv `notElem` fns]
-
-restrictRepositoryMapToGitRepoName:: RepositoryMap -> GitRepoName -> RepositoryMap
-restrictRepositoryMapToGitRepoName repoMap repoName = maybe M.empty (M.singleton repoName) $ repoName `M.lookup` repoMap
 
 uploadFrameworksAndDsymsToCaches :: RomeCacheInfo -> InvertedRepositoryMap -> [FrameworkVersion] -> [TargetPlatform] -> ReaderT UDCEnv IO ()
 uploadFrameworksAndDsymsToCaches cacheInfo reverseRomeMap fvs = mapM_ (sequence . uploadFramework)
@@ -257,7 +127,7 @@ uploadFrameworkAndDsymToCaches  (RomeCacheInfo bucketName localCacheDir) reverse
     dSYMNameWithDSYMExtension = frameworkNameWithFrameworkExtension ++ ".dSYM"
     dSYMdirectory = platformBuildDirectory </> dSYMNameWithDSYMExtension
     remoteDsymUploadPath = remoteDsymPath platform reverseRomeMap f version
-    zipDir dir verbose = liftIO $ Zip.addFilesToArchive (zipOptions False) Zip.emptyArchive [dir]
+    zipDir dir verbose = liftIO $ Zip.addFilesToArchive [Zip.OptRecursive] Zip.emptyArchive [dir]
 
 uploadBinary s3BucketName binaryZip destinationPath objectName = do
   (env, verbose) <- ask
@@ -269,7 +139,7 @@ uploadBinary s3BucketName binaryZip destinationPath objectName = do
       sayFunc $ "Started uploading " <> objectName <> " to: " <> destinationPath
     rs <- AWS.trying AWS._Error (AWS.send $ S3.putObject s3BucketName objectKey body)
     case rs of
-      Left e -> sayFunc $ "Error uploading " <> objectName <> ": " <> errorString e
+      Left e -> sayFunc $ "Error uploading " <> objectName <> ": " <> awsErrorToString e
       Right _ -> sayFunc $ "Uploaded " <> objectName <> " to: " <> destinationPath
 
 saveBinaryToLocalCache :: MonadIO m => FilePath -> LBS.ByteString -> FilePath -> String -> Bool -> m ()
@@ -299,7 +169,7 @@ downloadFrameworkAndDsymFromCaches (RomeCacheInfo bucketName localCacheDir) reve
       when skipLocalCache $ do
         eitherFrameworkBinary <- AWS.trying AWS._Error $ downloadBinary s3BucketName remoteFrameworkUploadPath fwn
         case eitherFrameworkBinary of
-          Left e -> sayFunc $ "Error downloading " <> fwn <> " : " <> errorString e
+          Left e -> sayFunc $ "Error downloading " <> fwn <> " : " <> awsErrorToString e
           Right frameworkBinary -> unzipBinary frameworkBinary fwn frameworkZipName verbose
 
       unless skipLocalCache $ do
@@ -313,7 +183,7 @@ downloadFrameworkAndDsymFromCaches (RomeCacheInfo bucketName localCacheDir) reve
         unless frameworkExistsInLocalCache $ do
           eitherFrameworkBinary <- AWS.trying AWS._Error $ downloadBinary s3BucketName remoteFrameworkUploadPath fwn
           case eitherFrameworkBinary of
-            Left e -> sayFunc $ "Error downloading " <> fwn <> " : " <> errorString e
+            Left e -> sayFunc $ "Error downloading " <> fwn <> " : " <> awsErrorToString e
             Right frameworkBinary -> do
               saveBinaryToLocalCache cacheDir frameworkBinary remoteFrameworkUploadPath fwn verbose
               unzipBinary frameworkBinary fwn frameworkZipName verbose
@@ -321,7 +191,7 @@ downloadFrameworkAndDsymFromCaches (RomeCacheInfo bucketName localCacheDir) reve
       when skipLocalCache $ do
         eitherdSYMBinary <- AWS.trying AWS._Error $ downloadBinary s3BucketName remotedSYMUploadPath dSYMName
         case eitherdSYMBinary of
-          Left e -> sayFunc $ "Error downloading " <> dSYMName <> " : " <> errorString e
+          Left e -> sayFunc $ "Error downloading " <> dSYMName <> " : " <> awsErrorToString e
           Right dSYMBinary -> unzipBinary dSYMBinary fwn dSYMZipName verbose
 
       unless skipLocalCache $ do
@@ -335,7 +205,7 @@ downloadFrameworkAndDsymFromCaches (RomeCacheInfo bucketName localCacheDir) reve
         unless dSYMExistsInLocalCache $ do
           eitherdSYMBinary <- AWS.trying AWS._Error $ downloadBinary s3BucketName remotedSYMUploadPath dSYMName
           case eitherdSYMBinary of
-            Left e -> sayFunc $ "Error downloading " <> dSYMName <> " : " <> errorString e
+            Left e -> sayFunc $ "Error downloading " <> dSYMName <> " : " <> awsErrorToString e
             Right dSYMBinary -> do
               saveBinaryToLocalCache cacheDir dSYMBinary remotedSYMUploadPath dSYMName verbose
               unzipBinary dSYMBinary fwn dSYMZipName verbose
@@ -343,12 +213,12 @@ downloadFrameworkAndDsymFromCaches (RomeCacheInfo bucketName localCacheDir) reve
     Nothing -> do
       eitherFrameworkBinary <- AWS.trying AWS._Error $ downloadBinary s3BucketName remoteFrameworkUploadPath fwn
       case eitherFrameworkBinary of
-        Left e -> sayFunc $ "Error downloading " <> fwn <> " : " <> errorString e
+        Left e -> sayFunc $ "Error downloading " <> fwn <> " : " <> awsErrorToString e
         Right frameworkBinary -> unzipBinary frameworkBinary fwn frameworkZipName verbose
 
       eitherdSYMBinary <- AWS.trying AWS._Error $ downloadBinary s3BucketName remotedSYMUploadPath (fwn ++ ".dSYM")
       case eitherdSYMBinary of
-        Left e -> sayFunc $ "Error downloading " <> dSYMName <> " : " <> errorString e
+        Left e -> sayFunc $ "Error downloading " <> dSYMName <> " : " <> awsErrorToString e
         Right dSYMBinary -> unzipBinary dSYMBinary fwn dSYMZipName verbose
 
   where
@@ -389,29 +259,13 @@ downloadBinary s3BucketName objectRemotePath objectName = do
             let a = if diffGreaterThan1MB then len else lastLen
             loop t len a)
 
-roundBytesToMegabytes :: Integral a => a -> Double
-roundBytesToMegabytes n = fromInteger (round (nInMB * (10^2))) / (10.0^^2)
-  where
-    nInMB = fromIntegral n / (1024*1024)
-
 unzipBinary :: MonadIO m => LBS.ByteString -> String -> String -> Bool -> m ()
 unzipBinary objectBinary objectName objectZipName verbose = do
   when verbose $
    sayLnWithTime $ "Staring to unzip " <> objectZipName
-  liftIO $ Zip.extractFilesFromArchive (zipOptions False) (Zip.toArchive objectBinary)
+  liftIO $ Zip.extractFilesFromArchive [Zip.OptRecursive] (Zip.toArchive objectBinary)
   when verbose $
     sayLnWithTime $ "Unzipped " <> objectName <> " from: " <> objectZipName
-
-remoteFrameworkPath :: TargetPlatform -> InvertedRepositoryMap -> FrameworkName -> Version -> String
-remoteFrameworkPath p r f v = remoteCacheDirectory p r f ++ frameworkArchiveName f v
-
-remoteDsymPath :: TargetPlatform -> InvertedRepositoryMap -> FrameworkName -> Version -> String
-remoteDsymPath p r f v = remoteCacheDirectory p r f ++ dSYMArchiveName f v
-
-remoteCacheDirectory :: TargetPlatform -> InvertedRepositoryMap -> FrameworkName -> String
-remoteCacheDirectory p r f = repoName </> show p ++ "/"
-  where
-    repoName = unGitRepoName $ repoNameForFrameworkName r f
 
 probeCachesForFrameworks :: RomeCacheInfo -> InvertedRepositoryMap -> [FrameworkVersion] -> [TargetPlatform] -> ReaderT (AWS.Env, Bool) IO [FrameworkAvailability]
 probeCachesForFrameworks cacheInfo reverseRomeMap frameworkVersions = sequence . probeForEachFramework
@@ -439,54 +293,6 @@ checkIfFrameworkExistsInBucket s3BucketName frameworkObjectKey verbose = do
   case rs of
     Left e -> return False
     Right hoResponse -> return True
-
-errorString :: AWS.Error -> String
-errorString e = fromErrorMessage $ fromMaybe (AWS.ErrorMessage "Unexpected Error") maybeServiceError
-  where
-    maybeServiceError = view AWS.serviceMessage =<< (e ^? AWS._ServiceError)
-
-sayLn :: MonadIO m => String -> m ()
-sayLn = liftIO . putStrLn
-
-sayLnWithTime :: MonadIO m => String -> m ()
-sayLnWithTime line = do
-  time <- liftIO getZonedTime
-  sayLn $ formatTime defaultTimeLocale "%T %F" time <> " - " <> line
-
-zipOptions :: Bool -> [Zip.ZipOption]
-zipOptions verbose = if verbose then [Zip.OptRecursive, Zip.OptVerbose] else [Zip.OptRecursive]
-
-deriveFrameworkNamesAndVersion :: RepositoryMap -> [CartfileEntry] -> [FrameworkVersion]
-deriveFrameworkNamesAndVersion romeMap = concatMap (deriveFrameworkNameAndVersion romeMap)
-
-deriveFrameworkNameAndVersion ::  RepositoryMap -> CartfileEntry -> [FrameworkVersion]
-deriveFrameworkNameAndVersion romeMap cfe@(CartfileEntry GitHub (Location l) v) = map (`FrameworkVersion` v) $
-  fromMaybe [FrameworkName gitHubRepositoryName] (M.lookup (gitRepoNameFromCartfileEntry cfe) romeMap)
-  where
-    gitHubRepositoryName = unGitRepoName $ gitRepoNameFromCartfileEntry cfe
-deriveFrameworkNameAndVersion romeMap cfe@(CartfileEntry Git (Location l) v)    = map (`FrameworkVersion` v) $
-  fromMaybe [FrameworkName gitRepositoryName] (M.lookup (gitRepoNameFromCartfileEntry cfe) romeMap)
-  where
-    gitRepositoryName = unGitRepoName $ gitRepoNameFromCartfileEntry cfe
-
-gitRepoNameFromCartfileEntry :: CartfileEntry -> GitRepoName
-gitRepoNameFromCartfileEntry (CartfileEntry GitHub (Location l) _) = GitRepoName . last . splitWithSeparator '/' $ l
-gitRepoNameFromCartfileEntry (CartfileEntry Git (Location l) _) = GitRepoName . replace ".git" "" . last . splitWithSeparator '/' $ l
-
-filterCartfileEntriesByGitRepoNames :: [GitRepoName] -> [CartfileEntry] -> [CartfileEntry]
-filterCartfileEntriesByGitRepoNames repoNames cartfileEntries = [c | c <- cartfileEntries, gitRepoNameFromCartfileEntry c `elem` repoNames]
-
-appendFrameworkExtensionTo :: FrameworkName -> String
-appendFrameworkExtensionTo (FrameworkName a) = a ++ ".framework"
-
-frameworkArchiveName :: FrameworkName -> Version -> String
-frameworkArchiveName f (Version v)  = appendFrameworkExtensionTo f ++ "-" ++ v ++ ".zip"
-
-dSYMArchiveName :: FrameworkName -> Version -> String
-dSYMArchiveName f (Version v) = appendFrameworkExtensionTo f ++ ".dSYM" ++ "-" ++ v ++ ".zip"
-
-splitWithSeparator :: Char -> String -> [String]
-splitWithSeparator a as = map T.unpack (T.split (== a) $ T.pack as)
 
 getMergedGitRepoAvailabilitiesFromFrameworkAvailabilities :: InvertedRepositoryMap -> [FrameworkAvailability] -> [GitRepoAvailability]
 getMergedGitRepoAvailabilitiesFromFrameworkAvailabilities reverseRomeMap = concatMap mergeRepoAvailabilities . groupAvailabilities . getGitRepoAvalabilities
@@ -516,27 +322,14 @@ formattedRepoAvailability listMode r@(GitRepoAvailability (GitRepoName rn) (Vers
     filteredAvailabilities = filterAccordingToListMode listMode pas
     formattedAvailabilities = unwords (formattedPlatformAvailability <$> filteredAvailabilities)
 
-
 filterAccordingToListMode :: ListMode -> [PlatformAvailability] -> [PlatformAvailability]
-filterAccordingToListMode All = id
-filterAccordingToListMode Missing = filter (not . _isAvailable)
-filterAccordingToListMode Present = filter _isAvailable
-
-formattedPlatformAvailability :: PlatformAvailability -> String
-formattedPlatformAvailability p = availabilityPrefix p ++ platformName p
-  where
-    availabilityPrefix (PlatformAvailability _ True) = "+"
-    availabilityPrefix (PlatformAvailability _ False) = "-"
-    platformName = show . _availabilityPlatform
-
-s3ConfigFile :: MonadIO m => m FilePath
-s3ConfigFile = (++ p) `liftM` liftIO getHomeDirectory
-  where
-      p = "/.aws/config"
+filterAccordingToListMode Commands.All     = id
+filterAccordingToListMode Commands.Missing = filter (not . _isAvailable)
+filterAccordingToListMode Commands.Present = filter _isAvailable
 
 discoverRegion :: RomeMonad AWS.Region
 discoverRegion = do
-  f <- s3ConfigFile
+  f <- getS3ConfigFile
   profile <- liftIO $ lookupEnv "AWS_PROFILE"
   getRegionFromFile f (fromMaybe "default" profile)
 
@@ -551,18 +344,3 @@ getRegionFromFile f profile = do
       case eitherAWSRegion of
         Left e  -> throwError e
         Right r -> return r
-
-toRomeFilesEntriesMap :: [RomefileEntry] -> RepositoryMap
-toRomeFilesEntriesMap = M.fromList . map romeFileEntryToTuple
-
-toInvertedRomeFilesEntriesMap :: [RomefileEntry] -> InvertedRepositoryMap
-toInvertedRomeFilesEntriesMap = M.fromList . concatMap romeFileEntryToListOfTuples
-  where listify (fs, g) = map (\f -> (f,g)) fs
-        flipTuple = uncurry (flip (,))
-        romeFileEntryToListOfTuples = listify . flipTuple . romeFileEntryToTuple
-
-romeFileEntryToTuple :: RomefileEntry -> (GitRepoName, [FrameworkName])
-romeFileEntryToTuple RomefileEntry {..} = (gitRepositoryName, frameworkCommonNames)
-
-repoNameForFrameworkName :: InvertedRepositoryMap -> FrameworkName -> GitRepoName
-repoNameForFrameworkName reverseRomeMap frameworkName = fromMaybe (GitRepoName . unFrameworkName $ frameworkName) (M.lookup frameworkName reverseRomeMap)
